Moved tests directory to test_framework to avoid package name conflicts
